+++ src/ArduinoJson/Polyfills/parseFloat.hpp
@@ -20,7 +20,7 @@ inline T parseFloat(const char* s) {
   typedef typename traits::mantissa_type mantissa_t;
   typedef typename traits::exponent_type exponent_t;
 
-  if (!s) return 0;
+  if (!s) return 0;  // NULL
 
   bool negative_result = false;
   switch (*s) {
@@ -30,6 +30,7 @@ inline T parseFloat(const char* s) {
       s++;
   }
 
+  if (*s == 't') return 1;  // true
   if (*s == 'n' || *s == 'N') return traits::nan();
   if (*s == 'i' || *s == 'I')
     return negative_result ? -traits::inf() : traits::inf();